fix: Resolve Kotlin compilation errors - Fix JoyStick.kt: Add Math.toDegrees for angle conversion - Fix Sprite.kt: Remove conflicting getter methods (getX, getY, getWidth, getHeight, getRadius)   - Changed protected var to public var properties   - Removed duplicate method declarations that conflicted with Kotlin properties - Fix MapLayer.kt: Correct Long/Int type conversions - Fix TiledBackground.kt: Correct Long/Int comparisons - Fix BitmapPool.kt: Add null safety (GameView.view!!) - Fix Sound.kt: Add null safety for context access - Fix Gauge.kt: Add null safety for resources access All compilation errors resolved for successful build.

+++ TuDefence/a2dg/src/main/kotlin/kr/ac/tukorea/ge/spgp2025/a2dg/framework/objects/TiledBackground.kt
@@ -112,9 +112,9 @@ class TiledBackground(mapAssetFile: String, var tileWidth: Float, var tileHeight
                 val drawCol = if (wraps) col % layer.width.toInt() else col
                 val drawRow = if (wraps) row % layer.height.toInt() else row
 
-                if (drawCol < 0 || drawCol >= layer.width || drawRow < 0 || drawRow >= layer.height) continue
+                if (drawCol < 0 || drawCol >= layer.width.toInt() || drawRow < 0 || drawRow >= layer.height.toInt()) continue
 
-                val tileId = layer.data!![(drawRow * layer.width.toInt() + drawCol).toLong()].toInt()
+                val tileId = layer.data!![(drawRow.toLong() * layer.width + drawCol.toLong()).toInt()].toInt()
                 if (tileId == 0) continue
 
                 val screenX = col * tileWidth - scrollX
